requirements.txt
lxml==4.9.3
Pillow==8.4.0
python-pptx==0.6.21
XlsxWriter==3.1.2
